pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Darya Konnova <dashustic@yandex.ru>", "Rishat Vergasov <Zvezda-Rishat1991@yandex.ru>", ]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.108.0"}
uvicorn = {extras = ["standart"], version = "^0.25.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
sqladmin = "^0.16.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
bcrypt = "4.0.1"
passlib = "^1.7.4"
aiosmtplib = "^3.0.1"
fastapi-filter = {extras = ["sqlalchemy"], version = "^1.1.0"}
fastapi-pagination = "^0.12.14"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
fastapi-limiter = "^0.1.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
